+++ test/factories/residents.rb
@@ -13,7 +13,6 @@ FactoryBot.define do
 
     trait :minor do
       is_minor { true }
-      age_of_minor { rand(1..20) }
     end
 
     trait :with_properties do